fix test setup
