+++ knexfile.js
@@ -14,6 +14,18 @@ module.exports = {
     }
   },
 
+  test: {
+    client: 'pg',
+    connection: 'postgres://localhost/palette_picker_test',
+    migrations: {
+      directory: './migrations'
+    },
+    seeds: {
+      directory: './seeds/test'
+    },
+    useNullAsDefault: true,
+  },
+
   production: {
     client: 'pg',
     connection: process.env.DATABASE_URL + `?ssl=true`,
